requirements/requirements.txt
grpcio-tools==1.16.1
googleapis-common-protos==1.5.5
python-daemon==2.2.0
mysqlclient==1.3.13
SQLAlchemy==1.2.9
pytest==4.0.2
python-dotenv==0.10.1
podder-task-base>=0.0.2,<0.1
